Add files via upload
Here we perform multiple sequence alignment, pairwise sequence alignment and use a heatmap to visualize the similarity of three VEGFR2 sequences, one that is the common splice variant and two that are alternatively spliced. Multiple sequence alignment is done to print out the alignment so that we can visualize the results. Pairwise sequence alignment and heatmap analyses are done to rigorously answers the scientific question: How similar are the alternative splice variants of the vascular endothelial growth factor receptor 2 (VEGFR2) gene (complete cds)?